fix: a map that can unmark a wall, and every number re-measured against it
Re-running the tables after the physics fix turned up a controller that had
stopped driving. SystematicCoverage parked on a kidney at 23% coverage with
twenty minutes of battery left, and had been doing so quietly.

Two causes. OccupancyMap could only ever gain walls -- mark_free refused to
overwrite one -- and walls are stamped from the estimated pose, so a drifting
estimate painted a few spurious ones in open water every minute until they
joined up. The frontier search walks non-wall cells, so it reported nothing
reachable while a third of the pool sat uncovered behind the fence. And the
controller's give-up budget counted a tick as a failure whether or not it was
covering new floor, then stopped for good when it ran out.

The hull now clears walls under itself, which is the one piece of evidence
strong enough to overrule them, and running out of patience backs up and turns
instead of parking. EvidenceMap had already learned the first of these for the
online planners.

That costs a documented result. The systematic docstring, the roadmap and the
dynamics notes all carried a measurement showing calibration improved the pose
estimate five-fold and halved coverage. The better the calibration, the sooner
the cage closed; that is all it was measuring. Over five seeds it now changes
neither, and all three places say so.

Everything else here is the re-measurement the physics fix needed. The planner
matrix, the fleet matrix, the partition fairness table, the truth-against-
odometry pair, the fleet scaling table, the oracle comparison, the greedy
horizon table, the ergodic and forecast and divergence numbers, the traced-
photo areas, the hundred-episode batch, the README summaries, the replay
figures and the notebook outputs. Some of the conclusions moved with them: the
fourth robot no longer makes the pool worse, it just stops making the team
better, and dirt_oracle now stays ahead of the baseline at every horizon while
covering a quarter of the floor.

+++ src/zimablue/pool/presets.py
@@ -55,7 +55,7 @@ def _arc(
     a0 = np.arctan2(start[1] - cy, start[0] - cx)
     a1 = np.arctan2(end[1] - cy, end[0] - cx)
     sweep = (a1 - a0) % (2 * np.pi) if ccw else -((a0 - a1) % (2 * np.pi))
-    t = np.linspace(a0, a0 + sweep, max(8, int(round(abs(sweep) * radius / step))))
+    t = np.linspace(a0, a0 + sweep, max(8, round(abs(sweep) * radius / step)))
     return np.column_stack([cx + radius * np.cos(t), cy + radius * np.sin(t)])
